Struct aws_sdk_codebuild::client::fluent_builders::StartBuild

pub struct StartBuild<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> StartBuild<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(self) -> Result<StartBuildOutput, SdkError<StartBuildError>> where
R::Policy: SmithyRetryPolicy<StartBuildInputOperationOutputAlias, StartBuildOutput, StartBuildError, StartBuildInputOperationRetryAlias>,

pub fn project_name(self, inp: impl Into<String>) -> Self

The name of the CodeBuild build project to start running a build.

pub fn set_project_name(self, input: Option<String>) -> Self

pub fn secondary_sources_override(self, inp: impl Into<ProjectSource>) -> Self

Appends an item to secondarySourcesOverride.

To override the contents of this collection use set_secondary_sources_override.

An array of ProjectSource objects.

pub fn set_secondary_sources_override(
self,
input: Option<Vec<ProjectSource>>
) -> Self

pub fn secondary_sources_version_override(
self,
inp: impl Into<ProjectSourceVersion>
) -> Self

Appends an item to secondarySourcesVersionOverride.

To override the contents of this collection use set_secondary_sources_version_override.

An array of ProjectSourceVersion objects that specify one or more versions of the project's secondary sources to be used for this build only.

pub fn set_secondary_sources_version_override(
self,
input: Option<Vec<ProjectSourceVersion>>
) -> Self

pub fn source_version(self, inp: impl Into<String>) -> Self

The version of the build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

pub fn set_source_version(self, input: Option<String>) -> Self

pub fn artifacts_override(self, inp: ProjectArtifacts) -> Self

Build output artifact settings that override, for this build only, the latest ones already defined in the build project.

pub fn set_artifacts_override(self, input: Option<ProjectArtifacts>) -> Self

pub fn secondary_artifacts_override(
self,
inp: impl Into<ProjectArtifacts>
) -> Self

Appends an item to secondaryArtifactsOverride.

To override the contents of this collection use set_secondary_artifacts_override.

An array of ProjectArtifacts objects.

pub fn set_secondary_artifacts_override(
self,
input: Option<Vec<ProjectArtifacts>>
) -> Self

pub fn environment_variables_override(
self,
inp: impl Into<EnvironmentVariable>
) -> Self

Appends an item to environmentVariablesOverride.

To override the contents of this collection use set_environment_variables_override.

A set of environment variables that overrides, for this build only, the latest ones already defined in the build project.

pub fn set_environment_variables_override(
self,
input: Option<Vec<EnvironmentVariable>>
) -> Self

pub fn source_type_override(self, inp: SourceType) -> Self

A source input type, for this build, that overrides the source input defined in the build project.

pub fn set_source_type_override(self, input: Option<SourceType>) -> Self

pub fn source_location_override(self, inp: impl Into<String>) -> Self

A location that overrides, for this build, the source location for the one defined in the build project.

pub fn set_source_location_override(self, input: Option<String>) -> Self

pub fn source_auth_override(self, inp: SourceAuth) -> Self

An authorization type for this build that overrides the one defined in the build project. This override applies only if the build project's source is BitBucket or GitHub.

pub fn set_source_auth_override(self, input: Option<SourceAuth>) -> Self

pub fn git_clone_depth_override(self, inp: i32) -> Self

The user-defined depth of history, with a minimum value of 0, that overrides, for this build only, any previous depth of history defined in the build project.

pub fn set_git_clone_depth_override(self, input: Option<i32>) -> Self

pub fn git_submodules_config_override(self, inp: GitSubmodulesConfig) -> Self

Information about the Git submodules configuration for this build of an CodeBuild build project.

pub fn set_git_submodules_config_override(
self,
input: Option<GitSubmodulesConfig>
) -> Self

pub fn buildspec_override(self, inp: impl Into<String>) -> Self

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

pub fn set_buildspec_override(self, input: Option<String>) -> Self

pub fn insecure_ssl_override(self, inp: bool) -> Self

Enable this flag to override the insecure SSL setting that is specified in the build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

pub fn set_insecure_ssl_override(self, input: Option<bool>) -> Self

pub fn report_build_status_override(self, inp: bool) -> Self

Set to true to report to your source provider the status of a build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

To be able to report the build status to the source provider, the user associated with the source provider must have write access to the repo. If the user does not have write access, the build status cannot be updated. For more information, see Source provider access in the CodeBuild User Guide.

The status of a build triggered by a webhook is always reported to your source provider.

pub fn set_report_build_status_override(self, input: Option<bool>) -> Self

pub fn build_status_config_override(self, inp: BuildStatusConfig) -> Self

Contains information that defines how the build project reports the build status to the source provider. This option is only used when the source provider is GITHUB, GITHUB_ENTERPRISE, or BITBUCKET.

pub fn set_build_status_config_override(
self,
input: Option<BuildStatusConfig>
) -> Self

pub fn environment_type_override(self, inp: EnvironmentType) -> Self

A container type for this build that overrides the one specified in the build project.

pub fn set_environment_type_override(
self,
input: Option<EnvironmentType>
) -> Self

pub fn image_override(self, inp: impl Into<String>) -> Self

The name of an image for this build that overrides the one specified in the build project.

pub fn set_image_override(self, input: Option<String>) -> Self

pub fn compute_type_override(self, inp: ComputeType) -> Self

The name of a compute type for this build that overrides the one specified in the build project.

pub fn set_compute_type_override(self, input: Option<ComputeType>) -> Self

pub fn certificate_override(self, inp: impl Into<String>) -> Self

The name of a certificate for this build that overrides the one specified in the build project.

pub fn set_certificate_override(self, input: Option<String>) -> Self

pub fn cache_override(self, inp: ProjectCache) -> Self

A ProjectCache object specified for this build that overrides the one defined in the build project.

pub fn set_cache_override(self, input: Option<ProjectCache>) -> Self

pub fn service_role_override(self, inp: impl Into<String>) -> Self

The name of a service role for this build that overrides the one specified in the build project.

pub fn set_service_role_override(self, input: Option<String>) -> Self

pub fn privileged_mode_override(self, inp: bool) -> Self

Enable this flag to override privileged mode in the build project.

pub fn set_privileged_mode_override(self, input: Option<bool>) -> Self

pub fn timeout_in_minutes_override(self, inp: i32) -> Self

The number of build timeout minutes, from 5 to 480 (8 hours), that overrides, for this build only, the latest setting already defined in the build project.

pub fn set_timeout_in_minutes_override(self, input: Option<i32>) -> Self

pub fn queued_timeout_in_minutes_override(self, inp: i32) -> Self

The number of minutes a build is allowed to be queued before it times out.

pub fn set_queued_timeout_in_minutes_override(self, input: Option<i32>) -> Self

pub fn encryption_key_override(self, inp: impl Into<String>) -> Self

The Key Management Service customer master key (CMK) that overrides the one specified in the build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/).

pub fn set_encryption_key_override(self, input: Option<String>) -> Self

pub fn idempotency_token(self, inp: impl Into<String>) -> Self

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuild request. The token is included in the StartBuild request and is valid for 5 minutes. If you repeat the StartBuild request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

pub fn set_idempotency_token(self, input: Option<String>) -> Self

pub fn logs_config_override(self, inp: LogsConfig) -> Self

Log settings for this build that override the log settings defined in the build project.

pub fn set_logs_config_override(self, input: Option<LogsConfig>) -> Self

pub fn registry_credential_override(self, inp: RegistryCredential) -> Self

The credentials for access to a private registry.

pub fn set_registry_credential_override(
self,
input: Option<RegistryCredential>
) -> Self

pub fn image_pull_credentials_type_override(
self,
inp: ImagePullCredentialsType
) -> Self

The type of credentials CodeBuild uses to pull images in your build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

pub fn set_image_pull_credentials_type_override(
self,
input: Option<ImagePullCredentialsType>
) -> Self

pub fn debug_session_enabled(self, inp: bool) -> Self

Specifies if session debugging is enabled for this build. For more information, see Viewing a running build in Session Manager.

pub fn set_debug_session_enabled(self, input: Option<bool>) -> Self