Struct aws_sdk_codebuild::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for CodeBuild_20161006.

This client allows ergonomic access to a CodeBuild_20161006-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_delete_builds(&self) -> BatchDeleteBuilds<C, M, R>

pub fn batch_get_build_batches(&self) -> BatchGetBuildBatches<C, M, R>

pub fn batch_get_builds(&self) -> BatchGetBuilds<C, M, R>

pub fn batch_get_projects(&self) -> BatchGetProjects<C, M, R>

pub fn batch_get_report_groups(&self) -> BatchGetReportGroups<C, M, R>

pub fn batch_get_reports(&self) -> BatchGetReports<C, M, R>

pub fn create_project(&self) -> CreateProject<C, M, R>

pub fn create_report_group(&self) -> CreateReportGroup<C, M, R>

pub fn create_webhook(&self) -> CreateWebhook<C, M, R>

pub fn delete_build_batch(&self) -> DeleteBuildBatch<C, M, R>

pub fn delete_project(&self) -> DeleteProject<C, M, R>

pub fn delete_report(&self) -> DeleteReport<C, M, R>

pub fn delete_report_group(&self) -> DeleteReportGroup<C, M, R>

pub fn delete_resource_policy(&self) -> DeleteResourcePolicy<C, M, R>

pub fn delete_source_credentials(&self) -> DeleteSourceCredentials<C, M, R>

pub fn delete_webhook(&self) -> DeleteWebhook<C, M, R>

pub fn describe_code_coverages(&self) -> DescribeCodeCoverages<C, M, R>

pub fn describe_test_cases(&self) -> DescribeTestCases<C, M, R>

pub fn get_report_group_trend(&self) -> GetReportGroupTrend<C, M, R>

pub fn get_resource_policy(&self) -> GetResourcePolicy<C, M, R>

pub fn import_source_credentials(&self) -> ImportSourceCredentials<C, M, R>

pub fn invalidate_project_cache(&self) -> InvalidateProjectCache<C, M, R>

pub fn list_build_batches(&self) -> ListBuildBatches<C, M, R>

pub fn list_build_batches_for_project(
&self
) -> ListBuildBatchesForProject<C, M, R>

pub fn list_builds(&self) -> ListBuilds<C, M, R>

pub fn list_builds_for_project(&self) -> ListBuildsForProject<C, M, R>

pub fn list_curated_environment_images(
&self
) -> ListCuratedEnvironmentImages<C, M, R>

pub fn list_projects(&self) -> ListProjects<C, M, R>

pub fn list_report_groups(&self) -> ListReportGroups<C, M, R>

pub fn list_reports(&self) -> ListReports<C, M, R>

pub fn list_reports_for_report_group(
&self
) -> ListReportsForReportGroup<C, M, R>

pub fn list_shared_projects(&self) -> ListSharedProjects<C, M, R>

pub fn list_shared_report_groups(&self) -> ListSharedReportGroups<C, M, R>

pub fn list_source_credentials(&self) -> ListSourceCredentials<C, M, R>

pub fn put_resource_policy(&self) -> PutResourcePolicy<C, M, R>

pub fn retry_build(&self) -> RetryBuild<C, M, R>

pub fn retry_build_batch(&self) -> RetryBuildBatch<C, M, R>

pub fn start_build(&self) -> StartBuild<C, M, R>

pub fn start_build_batch(&self) -> StartBuildBatch<C, M, R>

pub fn stop_build(&self) -> StopBuild<C, M, R>

pub fn stop_build_batch(&self) -> StopBuildBatch<C, M, R>

pub fn update_project(&self) -> UpdateProject<C, M, R>

pub fn update_project_visibility(&self) -> UpdateProjectVisibility<C, M, R>

pub fn update_report_group(&self) -> UpdateReportGroup<C, M, R>

pub fn update_webhook(&self) -> UpdateWebhook<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self