Module aws_sdk_codebuild::error

Modules

account_limit_exceeded_exception

See AccountLimitExceededException

invalid_input_exception

See InvalidInputException

o_auth_provider_exception

See OAuthProviderException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

Structs

AccountLimitExceededException

An Amazon Web Services service limit was exceeded for the calling Amazon Web Services account.

BatchDeleteBuildsError

BatchGetBuildBatchesError

BatchGetBuildsError

BatchGetProjectsError

BatchGetReportGroupsError

BatchGetReportsError

CreateProjectError

CreateReportGroupError

CreateWebhookError

DeleteBuildBatchError

DeleteProjectError

DeleteReportError

DeleteReportGroupError

DeleteResourcePolicyError

DeleteSourceCredentialsError

DeleteWebhookError

DescribeCodeCoveragesError

DescribeTestCasesError

GetReportGroupTrendError

GetResourcePolicyError

ImportSourceCredentialsError

InvalidInputException

The input value that was provided is not valid.

InvalidateProjectCacheError

ListBuildBatchesError

ListBuildBatchesForProjectError

ListBuildsError

ListBuildsForProjectError

ListCuratedEnvironmentImagesError

ListProjectsError

ListReportGroupsError

ListReportsError

ListReportsForReportGroupError

ListSharedProjectsError

ListSharedReportGroupsError

ListSourceCredentialsError

OAuthProviderException

There was a problem with the underlying OAuth provider.

PutResourcePolicyError

ResourceAlreadyExistsException

The specified Amazon Web Services resource cannot be created, because an Amazon Web Services resource with the same

ResourceNotFoundException

The specified Amazon Web Services resource cannot be found.

RetryBuildBatchError

RetryBuildError

StartBuildBatchError

StartBuildError

StopBuildBatchError

StopBuildError

UpdateProjectError

UpdateProjectVisibilityError

UpdateReportGroupError

UpdateWebhookError

Enums

BatchDeleteBuildsErrorKind

BatchGetBuildBatchesErrorKind

BatchGetBuildsErrorKind

BatchGetProjectsErrorKind

BatchGetReportGroupsErrorKind

BatchGetReportsErrorKind

CreateProjectErrorKind

CreateReportGroupErrorKind

CreateWebhookErrorKind

DeleteBuildBatchErrorKind

DeleteProjectErrorKind

DeleteReportErrorKind

DeleteReportGroupErrorKind

DeleteResourcePolicyErrorKind

DeleteSourceCredentialsErrorKind

DeleteWebhookErrorKind

DescribeCodeCoveragesErrorKind

DescribeTestCasesErrorKind

GetReportGroupTrendErrorKind

GetResourcePolicyErrorKind

ImportSourceCredentialsErrorKind

InvalidateProjectCacheErrorKind

ListBuildBatchesErrorKind

ListBuildBatchesForProjectErrorKind

ListBuildsErrorKind

ListBuildsForProjectErrorKind

ListCuratedEnvironmentImagesErrorKind

ListProjectsErrorKind

ListReportGroupsErrorKind

ListReportsErrorKind

ListReportsForReportGroupErrorKind

ListSharedProjectsErrorKind

ListSharedReportGroupsErrorKind

ListSourceCredentialsErrorKind

PutResourcePolicyErrorKind

RetryBuildBatchErrorKind

RetryBuildErrorKind

StartBuildBatchErrorKind

StartBuildErrorKind

StopBuildBatchErrorKind

StopBuildErrorKind

UpdateProjectErrorKind

UpdateProjectVisibilityErrorKind

UpdateReportGroupErrorKind

UpdateWebhookErrorKind