Struct aws_sdk_codebuild::input::StartBuildBatchInput

#[non_exhaustive]
pub struct StartBuildBatchInput {

    pub project_name: Option<String>,
    pub secondary_sources_override: Option<Vec<ProjectSource>>,
    pub secondary_sources_version_override: Option<Vec<ProjectSourceVersion>>,
    pub source_version: Option<String>,
    pub artifacts_override: Option<ProjectArtifacts>,
    pub secondary_artifacts_override: Option<Vec<ProjectArtifacts>>,
    pub environment_variables_override: Option<Vec<EnvironmentVariable>>,
    pub source_type_override: Option<SourceType>,
    pub source_location_override: Option<String>,
    pub source_auth_override: Option<SourceAuth>,
    pub git_clone_depth_override: Option<i32>,
    pub git_submodules_config_override: Option<GitSubmodulesConfig>,
    pub buildspec_override: Option<String>,
    pub insecure_ssl_override: Option<bool>,
    pub report_build_batch_status_override: Option<bool>,
    pub environment_type_override: Option<EnvironmentType>,
    pub image_override: Option<String>,
    pub compute_type_override: Option<ComputeType>,
    pub certificate_override: Option<String>,
    pub cache_override: Option<ProjectCache>,
    pub service_role_override: Option<String>,
    pub privileged_mode_override: Option<bool>,
    pub build_timeout_in_minutes_override: Option<i32>,
    pub queued_timeout_in_minutes_override: Option<i32>,
    pub encryption_key_override: Option<String>,
    pub idempotency_token: Option<String>,
    pub logs_config_override: Option<LogsConfig>,
    pub registry_credential_override: Option<RegistryCredential>,
    pub image_pull_credentials_type_override: Option<ImagePullCredentialsType>,
    pub build_batch_config_override: Option<ProjectBuildBatchConfig>,
    pub debug_session_enabled: Option<bool>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

project_name: Option<String>

The name of the project.

secondary_sources_override: Option<Vec<ProjectSource>>

An array of ProjectSource objects that override the secondary sources defined in the batch build project.

secondary_sources_version_override: Option<Vec<ProjectSourceVersion>>

An array of ProjectSourceVersion objects that override the secondary source versions in the batch build project.

source_version: Option<String>

The version of the batch build input to be built, for this build only. If not specified, the latest version is used. If specified, the contents depends on the source provider:

CodeCommit

The commit ID, branch, or Git tag to use.

GitHub

The commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Bitbucket

The commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

Amazon S3

The version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

artifacts_override: Option<ProjectArtifacts>

An array of ProjectArtifacts objects that contains information about the build output artifact overrides for the build project.

secondary_artifacts_override: Option<Vec<ProjectArtifacts>>

An array of ProjectArtifacts objects that override the secondary artifacts defined in the batch build project.

environment_variables_override: Option<Vec<EnvironmentVariable>>

An array of EnvironmentVariable objects that override, or add to, the environment variables defined in the batch build project.

source_type_override: Option<SourceType>

The source input type that overrides the source input defined in the batch build project.

source_location_override: Option<String>

A location that overrides, for this batch build, the source location defined in the batch build project.

source_auth_override: Option<SourceAuth>

A SourceAuth object that overrides the one defined in the batch build project. This override applies only if the build project's source is BitBucket or GitHub.

git_clone_depth_override: Option<i32>

The user-defined depth of history, with a minimum value of 0, that overrides, for this batch build only, any previous depth of history defined in the batch build project.

git_submodules_config_override: Option<GitSubmodulesConfig>

A GitSubmodulesConfig object that overrides the Git submodules configuration for this batch build.

buildspec_override: Option<String>

A buildspec file declaration that overrides, for this build only, the latest one already defined in the build project.

If this value is set, it can be either an inline buildspec definition, the path to an alternate buildspec file relative to the value of the built-in CODEBUILD_SRC_DIR environment variable, or the path to an S3 bucket. The bucket must be in the same Amazon Web Services Region as the build project. Specify the buildspec file using its ARN (for example, arn:aws:s3:::my-codebuild-sample2/buildspec.yml). If this value is not provided or is set to an empty string, the source code must contain a buildspec file in its root directory. For more information, see Buildspec File Name and Storage Location.

insecure_ssl_override: Option<bool>

Enable this flag to override the insecure SSL setting that is specified in the batch build project. The insecure SSL setting determines whether to ignore SSL warnings while connecting to the project source code. This override applies only if the build's source is GitHub Enterprise.

report_build_batch_status_override: Option<bool>

Set to true to report to your source provider the status of a batch build's start and completion. If you use this option with a source provider other than GitHub, GitHub Enterprise, or Bitbucket, an invalidInputException is thrown.

The status of a build triggered by a webhook is always reported to your source provider.

environment_type_override: Option<EnvironmentType>

A container type for this batch build that overrides the one specified in the batch build project.

image_override: Option<String>

The name of an image for this batch build that overrides the one specified in the batch build project.

compute_type_override: Option<ComputeType>

The name of a compute type for this batch build that overrides the one specified in the batch build project.

certificate_override: Option<String>

The name of a certificate for this batch build that overrides the one specified in the batch build project.

cache_override: Option<ProjectCache>

A ProjectCache object that specifies cache overrides.

service_role_override: Option<String>

The name of a service role for this batch build that overrides the one specified in the batch build project.

privileged_mode_override: Option<bool>

Enable this flag to override privileged mode in the batch build project.

build_timeout_in_minutes_override: Option<i32>

Overrides the build timeout specified in the batch build project.

queued_timeout_in_minutes_override: Option<i32>

The number of minutes a batch build is allowed to be queued before it times out.

encryption_key_override: Option<String>

The Key Management Service customer master key (CMK) that overrides the one specified in the batch build project. The CMK key encrypts the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/).

idempotency_token: Option<String>

A unique, case sensitive identifier you provide to ensure the idempotency of the StartBuildBatch request. The token is included in the StartBuildBatch request and is valid for five minutes. If you repeat the StartBuildBatch request with the same token, but change a parameter, CodeBuild returns a parameter mismatch error.

logs_config_override: Option<LogsConfig>

A LogsConfig object that override the log settings defined in the batch build project.

registry_credential_override: Option<RegistryCredential>

A RegistryCredential object that overrides credentials for access to a private registry.

image_pull_credentials_type_override: Option<ImagePullCredentialsType>

The type of credentials CodeBuild uses to pull images in your batch build. There are two valid values:

CODEBUILD

Specifies that CodeBuild uses its own credentials. This requires that you modify your ECR repository policy to trust CodeBuild's service principal.

SERVICE_ROLE

Specifies that CodeBuild uses your build project's service role.

When using a cross-account or private registry image, you must use SERVICE_ROLE credentials. When using an CodeBuild curated image, you must use CODEBUILD credentials.

build_batch_config_override: Option<ProjectBuildBatchConfig>

A BuildBatchConfigOverride object that contains batch build configuration overrides.

debug_session_enabled: Option<bool>

Specifies if session debugging is enabled for this batch build. For more information, see Viewing a running build in Session Manager. Batch session debugging is not supported for matrix batch builds.

Implementations

impl StartBuildBatchInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<StartBuildBatch, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<StartBuildBatch>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture StartBuildBatchInput