Module aws_sdk_codebuild::model

Modules

batch_restrictions

See BatchRestrictions

build

See Build

build_artifacts

See BuildArtifacts

build_batch

See BuildBatch

build_batch_filter

See BuildBatchFilter

build_batch_phase

See BuildBatchPhase

build_group

See BuildGroup

build_not_deleted

See BuildNotDeleted

build_phase

See BuildPhase

build_status_config

See BuildStatusConfig

build_summary

See BuildSummary

cloud_watch_logs_config

See CloudWatchLogsConfig

code_coverage

See CodeCoverage

code_coverage_report_summary

See CodeCoverageReportSummary

debug_session

See DebugSession

environment_image

See EnvironmentImage

environment_language

See EnvironmentLanguage

environment_platform

See EnvironmentPlatform

environment_variable

See EnvironmentVariable

exported_environment_variable

See ExportedEnvironmentVariable

git_submodules_config

See GitSubmodulesConfig

logs_config

See LogsConfig

logs_location

See LogsLocation

network_interface

See NetworkInterface

phase_context

See PhaseContext

project

See Project

project_artifacts

See ProjectArtifacts

project_badge

See ProjectBadge

project_build_batch_config

See ProjectBuildBatchConfig

project_cache

See ProjectCache

project_environment

See ProjectEnvironment

project_file_system_location

See ProjectFileSystemLocation

project_source

See ProjectSource

project_source_version

See ProjectSourceVersion

registry_credential

See RegistryCredential

report

See Report

report_export_config

See ReportExportConfig

report_filter

See ReportFilter

report_group

See ReportGroup

report_group_trend_stats

See ReportGroupTrendStats

report_with_raw_data

See ReportWithRawData

resolved_artifact

See ResolvedArtifact

s3_logs_config

See S3LogsConfig

s3_report_export_config

See S3ReportExportConfig

source_auth

See SourceAuth

source_credentials_info

See SourceCredentialsInfo

tag

See Tag

test_case

See TestCase

test_case_filter

See TestCaseFilter

test_report_summary

See TestReportSummary

vpc_config

See VpcConfig

webhook

See Webhook

webhook_filter

See WebhookFilter

Structs

BatchRestrictions

Specifies restrictions for the batch build.

Build

Information about a build.

BuildArtifacts

Information about build output artifacts.

BuildBatch

Contains information about a batch build.

BuildBatchFilter

Specifies filters when retrieving batch builds.

BuildBatchPhase

Contains information about a stage for a batch build.

BuildGroup

Contains information about a batch build build group. Build groups are used to combine

BuildNotDeleted

Information about a build that could not be successfully deleted.

BuildPhase

Information about a stage for a build.

BuildStatusConfig

Contains information that defines how the CodeBuild build project reports the build status

BuildSummary

Contains summary information about a batch build group.

CloudWatchLogsConfig

Information about CloudWatch Logs for a build project.

CodeCoverage

Contains code coverage report information.

CodeCoverageReportSummary

Contains a summary of a code coverage report.

DebugSession

Contains information about the debug session for a build. For more information, see Viewing a

EnvironmentImage

Information about a Docker image that is managed by CodeBuild.

EnvironmentLanguage

A set of Docker images that are related by programming language and are managed by

EnvironmentPlatform

A set of Docker images that are related by platform and are managed by CodeBuild.

EnvironmentVariable

Information about an environment variable for a build project or a build.

ExportedEnvironmentVariable

Contains information about an exported environment variable.

GitSubmodulesConfig

Information about the Git submodules configuration for an CodeBuild build project.

LogsConfig

Information about logs for a build project. These can be logs in CloudWatch Logs, built in a

LogsLocation

Information about build logs in CloudWatch Logs.

NetworkInterface

Describes a network interface.

PhaseContext

Additional information about a build phase that has an error. You can use this

Project

Information about a build project.

ProjectArtifacts

Information about the build output artifacts for the build project.

ProjectBadge

Information about the build badge for the build project.

ProjectBuildBatchConfig

Contains configuration information about a batch build project.

ProjectCache

Information about the cache for the build project.

ProjectEnvironment

Information about the build environment of the build project.

ProjectFileSystemLocation

Information about a file system created by Amazon Elastic File System (EFS). For more

ProjectSource

Information about the build input source code for the build project.

ProjectSourceVersion

A source identifier and its corresponding version.

RegistryCredential

Information about credentials that provide access to a private Docker registry. When

Report

Information about the results from running a series of test cases during the run of a

ReportExportConfig

Information about the location where the run of a report is exported.

ReportFilter

A filter used to return reports with the status specified by the input

ReportGroup

A series of reports. Each report contains information about the results from running a

ReportGroupTrendStats

Contains trend statistics for a set of reports. The actual values depend on the type of

ReportWithRawData

Contains the unmodified data for the report. For more information, see .

ResolvedArtifact

Represents a resolved build artifact. A resolved artifact is an artifact that is built and

S3LogsConfig

Information about S3 logs for a build project.

S3ReportExportConfig

Information about the S3 bucket where the raw data of a report are exported.

SourceAuth

Information about the authorization settings for CodeBuild to access the source code to be

SourceCredentialsInfo

Information about the credentials for a GitHub, GitHub Enterprise, or Bitbucket

Tag

A tag, consisting of a key and a value.

TestCase

Information about a test case created using a framework such as NUnit or Cucumber. A

TestCaseFilter

A filter used to return specific types of test cases. In order to pass the filter, the

TestReportSummary

Information about a test report.

VpcConfig

Information about the VPC configuration that CodeBuild accesses.

Webhook

Information about a webhook that connects repository events to a build project in

WebhookFilter

A filter used to determine which webhooks trigger a build.

Enums

ArtifactNamespace

ArtifactPackaging

ArtifactsType

AuthType

BucketOwnerAccess

Specifies the bucket owner's access for objects that another account uploads to their

BuildBatchPhaseType

BuildPhaseType

CacheMode

CacheType

ComputeType

CredentialProviderType

EnvironmentType

EnvironmentVariableType

FileSystemType

ImagePullCredentialsType

LanguageType

LogsConfigStatusType

PlatformType

ProjectSortByType

ProjectVisibilityType

Specifies the visibility of the project's builds. Possible values are:

ReportCodeCoverageSortByType

ReportExportConfigType

ReportGroupSortByType

ReportGroupStatusType

ReportGroupTrendFieldType

ReportPackagingType

ReportStatusType

ReportType

RetryBuildBatchType

ServerType

SharedResourceSortByType

SortOrderType

SourceAuthType

SourceType

StatusType

WebhookBuildType

WebhookFilterType