Struct aws_sdk_codebuild::model::project::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Project

Implementations

impl Builder

pub fn name(self, input: impl Into<String>) -> Self

The name of the build project.

pub fn set_name(self, input: Option<String>) -> Self

pub fn arn(self, input: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the build project.

pub fn set_arn(self, input: Option<String>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A description that makes the build project easy to identify.

pub fn set_description(self, input: Option<String>) -> Self

pub fn source(self, input: ProjectSource) -> Self

Information about the build input source code for this build project.

pub fn set_source(self, input: Option<ProjectSource>) -> Self

pub fn secondary_sources(self, input: impl Into<ProjectSource>) -> Self

pub fn set_secondary_sources(self, input: Option<Vec<ProjectSource>>) -> Self

pub fn source_version(self, input: impl Into<String>) -> Self

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

pub fn set_source_version(self, input: Option<String>) -> Self

pub fn secondary_source_versions(
self,
input: impl Into<ProjectSourceVersion>
) -> Self

pub fn set_secondary_source_versions(
self,
input: Option<Vec<ProjectSourceVersion>>
) -> Self

pub fn artifacts(self, input: ProjectArtifacts) -> Self

Information about the build output artifacts for the build project.

pub fn set_artifacts(self, input: Option<ProjectArtifacts>) -> Self

pub fn secondary_artifacts(self, input: impl Into<ProjectArtifacts>) -> Self

pub fn set_secondary_artifacts(
self,
input: Option<Vec<ProjectArtifacts>>
) -> Self

pub fn cache(self, input: ProjectCache) -> Self

Information about the cache for the build project.

pub fn set_cache(self, input: Option<ProjectCache>) -> Self

pub fn environment(self, input: ProjectEnvironment) -> Self

Information about the build environment for this build project.

pub fn set_environment(self, input: Option<ProjectEnvironment>) -> Self

pub fn service_role(self, input: impl Into<String>) -> Self

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

pub fn set_service_role(self, input: Option<String>) -> Self

pub fn timeout_in_minutes(self, input: i32) -> Self

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

pub fn set_timeout_in_minutes(self, input: Option<i32>) -> Self

pub fn queued_timeout_in_minutes(self, input: i32) -> Self

The number of minutes a build is allowed to be queued before it times out.

pub fn set_queued_timeout_in_minutes(self, input: Option<i32>) -> Self

pub fn encryption_key(self, input: impl Into<String>) -> Self

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/). If you don't specify a value, CodeBuild uses the managed CMK for Amazon Simple Storage Service (Amazon S3).

pub fn set_encryption_key(self, input: Option<String>) -> Self

pub fn tags(self, input: impl Into<Tag>) -> Self

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn created(self, input: Instant) -> Self

When the build project was created, expressed in Unix time format.

pub fn set_created(self, input: Option<Instant>) -> Self

pub fn last_modified(self, input: Instant) -> Self

When the build project's settings were last modified, expressed in Unix time format.

pub fn set_last_modified(self, input: Option<Instant>) -> Self

pub fn webhook(self, input: Webhook) -> Self

Information about a webhook that connects repository events to a build project in CodeBuild.

pub fn set_webhook(self, input: Option<Webhook>) -> Self

pub fn vpc_config(self, input: VpcConfig) -> Self

Information about the VPC configuration that CodeBuild accesses.

pub fn set_vpc_config(self, input: Option<VpcConfig>) -> Self

pub fn badge(self, input: ProjectBadge) -> Self

Information about the build badge for the build project.

pub fn set_badge(self, input: Option<ProjectBadge>) -> Self

pub fn logs_config(self, input: LogsConfig) -> Self

Information about logs for the build project. A project can create logs in CloudWatch Logs, an S3 bucket, or both.

pub fn set_logs_config(self, input: Option<LogsConfig>) -> Self

pub fn file_system_locations(
self,
input: impl Into<ProjectFileSystemLocation>
) -> Self

pub fn set_file_system_locations(
self,
input: Option<Vec<ProjectFileSystemLocation>>
) -> Self

pub fn build_batch_config(self, input: ProjectBuildBatchConfig) -> Self

A ProjectBuildBatchConfig object that defines the batch build options for the project.

pub fn set_build_batch_config(
self,
input: Option<ProjectBuildBatchConfig>
) -> Self

pub fn concurrent_build_limit(self, input: i32) -> Self

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

pub fn set_concurrent_build_limit(self, input: Option<i32>) -> Self

pub fn project_visibility(self, input: ProjectVisibilityType) -> Self

Specifies the visibility of the project's builds. Possible values are:

PUBLIC_READ

The project builds are visible to the public.

PRIVATE

The project builds are not visible to the public.

pub fn set_project_visibility(
self,
input: Option<ProjectVisibilityType>
) -> Self

pub fn public_project_alias(self, input: impl Into<String>) -> Self

Contains the project identifier used with the public build APIs.

pub fn set_public_project_alias(self, input: Option<String>) -> Self

pub fn resource_access_role(self, input: impl Into<String>) -> Self

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

pub fn set_resource_access_role(self, input: Option<String>) -> Self

pub fn build(self) -> Project

Consumes the builder and constructs a Project