Struct aws_sdk_codebuild::model::Build

#[non_exhaustive]
pub struct Build {

    pub id: Option<String>,
    pub arn: Option<String>,
    pub build_number: Option<i64>,
    pub start_time: Option<Instant>,
    pub end_time: Option<Instant>,
    pub current_phase: Option<String>,
    pub build_status: Option<StatusType>,
    pub source_version: Option<String>,
    pub resolved_source_version: Option<String>,
    pub project_name: Option<String>,
    pub phases: Option<Vec<BuildPhase>>,
    pub source: Option<ProjectSource>,
    pub secondary_sources: Option<Vec<ProjectSource>>,
    pub secondary_source_versions: Option<Vec<ProjectSourceVersion>>,
    pub artifacts: Option<BuildArtifacts>,
    pub secondary_artifacts: Option<Vec<BuildArtifacts>>,
    pub cache: Option<ProjectCache>,
    pub environment: Option<ProjectEnvironment>,
    pub service_role: Option<String>,
    pub logs: Option<LogsLocation>,
    pub timeout_in_minutes: Option<i32>,
    pub queued_timeout_in_minutes: Option<i32>,
    pub build_complete: bool,
    pub initiator: Option<String>,
    pub vpc_config: Option<VpcConfig>,
    pub network_interface: Option<NetworkInterface>,
    pub encryption_key: Option<String>,
    pub exported_environment_variables: Option<Vec<ExportedEnvironmentVariable>>,
    pub report_arns: Option<Vec<String>>,
    pub file_system_locations: Option<Vec<ProjectFileSystemLocation>>,
    pub debug_session: Option<DebugSession>,
    pub build_batch_arn: Option<String>,

}

Information about a build.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: Option<String>

The unique ID for the build.

arn: Option<String>

The Amazon Resource Name (ARN) of the build.

build_number: Option<i64>

The number of the build. For each project, the buildNumber of its first build is 1. The buildNumber of each subsequent build is incremented by 1. If a build is deleted, the buildNumber of other builds does not change.

start_time: Option<Instant>

When the build process started, expressed in Unix time format.

end_time: Option<Instant>

When the build process ended, expressed in Unix time format.

current_phase: Option<String>

The current build phase.

build_status: Option<StatusType>

The current status of the build. Valid values include:

FAILED: The build failed.

FAULT: The build faulted.

IN_PROGRESS: The build is still in progress.

STOPPED: The build stopped.

SUCCEEDED: The build succeeded.

TIMED_OUT: The build timed out.

source_version: Option<String>

Any version identifier for the version of the source code to be built. If sourceVersion is specified at the project level, then this sourceVersion (at the build level) takes precedence.

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

resolved_source_version: Option<String>

An identifier for the version of this build's source code.

For CodeCommit, GitHub, GitHub Enterprise, and BitBucket, the commit ID.

For CodePipeline, the source revision provided by CodePipeline.

For Amazon S3, this does not apply.

project_name: Option<String>

The name of the CodeBuild project.

phases: Option<Vec<BuildPhase>>

Information about all previous build phases that are complete and information about any current build phase that is not yet complete.

source: Option<ProjectSource>

Information about the source code to be built.

secondary_sources: Option<Vec<ProjectSource>>

An array of ProjectSource objects.

secondary_source_versions: Option<Vec<ProjectSourceVersion>>

An array of ProjectSourceVersion objects. Each ProjectSourceVersion must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example, pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

artifacts: Option<BuildArtifacts>

Information about the output artifacts for the build.

secondary_artifacts: Option<Vec<BuildArtifacts>>

An array of ProjectArtifacts objects.

cache: Option<ProjectCache>

Information about the cache for the build.

environment: Option<ProjectEnvironment>

Information about the build environment for this build.

service_role: Option<String>

The name of a service role used for this build.

logs: Option<LogsLocation>

Information about the build's logs in CloudWatch Logs.

timeout_in_minutes: Option<i32>

How long, in minutes, for CodeBuild to wait before timing out this build if it does not get marked as completed.

queued_timeout_in_minutes: Option<i32>

The number of minutes a build is allowed to be queued before it times out.

build_complete: bool

Whether the build is complete. True if complete; otherwise, false.

initiator: Option<String>

The entity that started the build. Valid values include:

If CodePipeline started the build, the pipeline's name (for example, codepipeline/my-demo-pipeline).

If an IAM user started the build, the user's name (for example, MyUserName).

If the Jenkins plugin for CodeBuild started the build, the string CodeBuild-Jenkins-Plugin.

vpc_config: Option<VpcConfig>

If your CodeBuild project accesses resources in an Amazon VPC, you provide this parameter that identifies the VPC ID and the list of security group IDs and subnet IDs. The security groups and subnets must belong to the same VPC. You must provide at least one security group and one subnet ID.

network_interface: Option<NetworkInterface>

Describes a network interface.

encryption_key: Option<String>

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/).

exported_environment_variables: Option<Vec<ExportedEnvironmentVariable>>

A list of exported environment variables for this build.

Exported environment variables are used in conjunction with CodePipeline to export environment variables from the current build stage to subsequent stages in the pipeline. For more information, see Working with variables in the CodePipeline User Guide.

report_arns: Option<Vec<String>>

An array of the ARNs associated with this build's reports.

file_system_locations: Option<Vec<ProjectFileSystemLocation>>

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

debug_session: Option<DebugSession>

Contains information about the debug session for this build.

build_batch_arn: Option<String>

The ARN of the batch build that this build is a member of, if applicable.

Implementations

impl Build

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Build