Struct aws_sdk_codebuild::model::Project

#[non_exhaustive]
pub struct Project {

    pub name: Option<String>,
    pub arn: Option<String>,
    pub description: Option<String>,
    pub source: Option<ProjectSource>,
    pub secondary_sources: Option<Vec<ProjectSource>>,
    pub source_version: Option<String>,
    pub secondary_source_versions: Option<Vec<ProjectSourceVersion>>,
    pub artifacts: Option<ProjectArtifacts>,
    pub secondary_artifacts: Option<Vec<ProjectArtifacts>>,
    pub cache: Option<ProjectCache>,
    pub environment: Option<ProjectEnvironment>,
    pub service_role: Option<String>,
    pub timeout_in_minutes: Option<i32>,
    pub queued_timeout_in_minutes: Option<i32>,
    pub encryption_key: Option<String>,
    pub tags: Option<Vec<Tag>>,
    pub created: Option<Instant>,
    pub last_modified: Option<Instant>,
    pub webhook: Option<Webhook>,
    pub vpc_config: Option<VpcConfig>,
    pub badge: Option<ProjectBadge>,
    pub logs_config: Option<LogsConfig>,
    pub file_system_locations: Option<Vec<ProjectFileSystemLocation>>,
    pub build_batch_config: Option<ProjectBuildBatchConfig>,
    pub concurrent_build_limit: Option<i32>,
    pub project_visibility: Option<ProjectVisibilityType>,
    pub public_project_alias: Option<String>,
    pub resource_access_role: Option<String>,

}

Information about a build project.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

name: Option<String>

The name of the build project.

arn: Option<String>

The Amazon Resource Name (ARN) of the build project.

description: Option<String>

A description that makes the build project easy to identify.

source: Option<ProjectSource>

Information about the build input source code for this build project.

secondary_sources: Option<Vec<ProjectSource>>

An array of ProjectSource objects.

source_version: Option<String>

A version of the build input to be built for this project. If not specified, the latest version is used. If specified, it must be one of:

For CodeCommit: the commit ID, branch, or Git tag to use.

For GitHub: the commit ID, pull request ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a pull request ID is specified, it must use the format pr/pull-request-ID (for example pr/25). If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Bitbucket: the commit ID, branch name, or tag name that corresponds to the version of the source code you want to build. If a branch name is specified, the branch's HEAD commit ID is used. If not specified, the default branch's HEAD commit ID is used.

For Amazon S3: the version ID of the object that represents the build input ZIP file to use.

If sourceVersion is specified at the build level, then that version takes precedence over this sourceVersion (at the project level).

For more information, see Source Version Sample with CodeBuild in the CodeBuild User Guide.

secondary_source_versions: Option<Vec<ProjectSourceVersion>>

An array of ProjectSourceVersion objects. If secondarySourceVersions is specified at the build level, then they take over these secondarySourceVersions (at the project level).

artifacts: Option<ProjectArtifacts>

Information about the build output artifacts for the build project.

secondary_artifacts: Option<Vec<ProjectArtifacts>>

An array of ProjectArtifacts objects.

cache: Option<ProjectCache>

Information about the cache for the build project.

environment: Option<ProjectEnvironment>

Information about the build environment for this build project.

service_role: Option<String>

The ARN of the IAM role that enables CodeBuild to interact with dependent Amazon Web Services services on behalf of the Amazon Web Services account.

timeout_in_minutes: Option<i32>

How long, in minutes, from 5 to 480 (8 hours), for CodeBuild to wait before timing out any related build that did not get marked as completed. The default is 60 minutes.

queued_timeout_in_minutes: Option<i32>

The number of minutes a build is allowed to be queued before it times out.

encryption_key: Option<String>

The Key Management Service customer master key (CMK) to be used for encrypting the build output artifacts.

You can use a cross-account KMS key to encrypt the build output artifacts if your service role has permission to that key.

You can specify either the Amazon Resource Name (ARN) of the CMK or, if available, the CMK's alias (using the format alias/). If you don't specify a value, CodeBuild uses the managed CMK for Amazon Simple Storage Service (Amazon S3).

tags: Option<Vec<Tag>>

A list of tag key and value pairs associated with this build project.

These tags are available for use by Amazon Web Services services that support CodeBuild build project tags.

created: Option<Instant>

When the build project was created, expressed in Unix time format.

last_modified: Option<Instant>

When the build project's settings were last modified, expressed in Unix time format.

webhook: Option<Webhook>

Information about a webhook that connects repository events to a build project in CodeBuild.

vpc_config: Option<VpcConfig>

Information about the VPC configuration that CodeBuild accesses.

badge: Option<ProjectBadge>

Information about the build badge for the build project.

logs_config: Option<LogsConfig>

Information about logs for the build project. A project can create logs in CloudWatch Logs, an S3 bucket, or both.

file_system_locations: Option<Vec<ProjectFileSystemLocation>>

An array of ProjectFileSystemLocation objects for a CodeBuild build project. A ProjectFileSystemLocation object specifies the identifier, location, mountOptions, mountPoint, and type of a file system created using Amazon Elastic File System.

build_batch_config: Option<ProjectBuildBatchConfig>

A ProjectBuildBatchConfig object that defines the batch build options for the project.

concurrent_build_limit: Option<i32>

The maximum number of concurrent builds that are allowed for this project.

New builds are only started if the current number of builds is less than or equal to this limit. If the current build count meets this limit, new builds are throttled and are not run.

project_visibility: Option<ProjectVisibilityType>

Specifies the visibility of the project's builds. Possible values are:

PUBLIC_READ

The project builds are visible to the public.

PRIVATE

The project builds are not visible to the public.

public_project_alias: Option<String>

Contains the project identifier used with the public build APIs.

resource_access_role: Option<String>

The ARN of the IAM role that enables CodeBuild to access the CloudWatch Logs and Amazon S3 artifacts for the project's builds.

Implementations

impl Project

pub fn builder() -> Builder

Creates a new builder-style object to manufacture Project