Module aws_sdk_codebuild::operation

Structs

BatchDeleteBuilds

Deletes one or more builds.

BatchGetBuildBatches

Retrieves information about one or more batch builds.

BatchGetBuilds

Gets information about one or more builds.

BatchGetProjects

Gets information about one or more build projects.

BatchGetReportGroups

BatchGetReports

CreateProject

Creates a build project.

CreateReportGroup

CreateWebhook

For an existing CodeBuild build project that has its source code stored in a GitHub or

DeleteBuildBatch

Deletes a batch build.

DeleteProject

Deletes a build project. When you delete a project, its builds are not deleted.

DeleteReport

DeleteReportGroup

Deletes a report group. Before you delete a report group, you must delete its reports.

DeleteResourcePolicy

Deletes a resource policy that is identified by its resource ARN.

DeleteSourceCredentials

Deletes a set of GitHub, GitHub Enterprise, or Bitbucket source credentials.

DeleteWebhook

For an existing CodeBuild build project that has its source code stored in a GitHub or

DescribeCodeCoverages

Retrieves one or more code coverage reports.

DescribeTestCases

GetReportGroupTrend

Analyzes and accumulates test report values for the specified test reports.

GetResourcePolicy

Gets a resource policy that is identified by its resource ARN.

ImportSourceCredentials

Imports the source repository credentials for an CodeBuild project that has its

InvalidateProjectCache

Resets the cache for a project.

ListBuildBatches

Retrieves the identifiers of your build batches in the current region.

ListBuildBatchesForProject

Retrieves the identifiers of the build batches for a specific project.

ListBuilds

Gets a list of build IDs, with each build ID representing a single build.

ListBuildsForProject

Gets a list of build identifiers for the specified build project, with each build

ListCuratedEnvironmentImages

Gets information about Docker images that are managed by CodeBuild.

ListProjects

Gets a list of build project names, with each build project name representing a single

ListReportGroups

ListReports

ListReportsForReportGroup

ListSharedProjects

Gets a list of projects that are shared with other Amazon Web Services accounts or users.

ListSharedReportGroups

Gets a list of report groups that are shared with other Amazon Web Services accounts or users.

ListSourceCredentials

Returns a list of SourceCredentialsInfo objects.

PutResourcePolicy

Stores a resource policy for the ARN of a Project or

RetryBuild

Restarts a build.

RetryBuildBatch

Restarts a failed batch build. Only batch builds that have failed can be retried.

StartBuild

Starts running a build.

StartBuildBatch

Starts a batch build for a project.

StopBuild

Attempts to stop running a build.

StopBuildBatch

Stops a running batch build.

UpdateProject

Changes the settings of a build project.

UpdateProjectVisibility

Changes the public visibility for a project. The project's build results, logs, and

UpdateReportGroup

UpdateWebhook

Updates the webhook associated with an CodeBuild build project.