Struct aws_sdk_codebuild::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for CodeBuild_20161006.

This client allows ergonomic access to a CodeBuild_20161006-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn batch_delete_builds(&self) -> BatchDeleteBuilds<C, M, R>

Constructs a fluent builder for the BatchDeleteBuilds operation.

See BatchDeleteBuilds for more information about the operation and its arguments.

pub fn batch_get_build_batches(&self) -> BatchGetBuildBatches<C, M, R>

Constructs a fluent builder for the BatchGetBuildBatches operation.

See BatchGetBuildBatches for more information about the operation and its arguments.

pub fn batch_get_builds(&self) -> BatchGetBuilds<C, M, R>

Constructs a fluent builder for the BatchGetBuilds operation.

See BatchGetBuilds for more information about the operation and its arguments.

pub fn batch_get_projects(&self) -> BatchGetProjects<C, M, R>

Constructs a fluent builder for the BatchGetProjects operation.

See BatchGetProjects for more information about the operation and its arguments.

pub fn batch_get_report_groups(&self) -> BatchGetReportGroups<C, M, R>

Constructs a fluent builder for the BatchGetReportGroups operation.

See BatchGetReportGroups for more information about the operation and its arguments.

pub fn batch_get_reports(&self) -> BatchGetReports<C, M, R>

Constructs a fluent builder for the BatchGetReports operation.

See BatchGetReports for more information about the operation and its arguments.

pub fn create_project(&self) -> CreateProject<C, M, R>

Constructs a fluent builder for the CreateProject operation.

See CreateProject for more information about the operation and its arguments.

pub fn create_report_group(&self) -> CreateReportGroup<C, M, R>

Constructs a fluent builder for the CreateReportGroup operation.

See CreateReportGroup for more information about the operation and its arguments.

pub fn create_webhook(&self) -> CreateWebhook<C, M, R>

Constructs a fluent builder for the CreateWebhook operation.

See CreateWebhook for more information about the operation and its arguments.

pub fn delete_build_batch(&self) -> DeleteBuildBatch<C, M, R>

Constructs a fluent builder for the DeleteBuildBatch operation.

See DeleteBuildBatch for more information about the operation and its arguments.

pub fn delete_project(&self) -> DeleteProject<C, M, R>

Constructs a fluent builder for the DeleteProject operation.

See DeleteProject for more information about the operation and its arguments.

pub fn delete_report(&self) -> DeleteReport<C, M, R>

Constructs a fluent builder for the DeleteReport operation.

See DeleteReport for more information about the operation and its arguments.

pub fn delete_report_group(&self) -> DeleteReportGroup<C, M, R>

Constructs a fluent builder for the DeleteReportGroup operation.

See DeleteReportGroup for more information about the operation and its arguments.

pub fn delete_resource_policy(&self) -> DeleteResourcePolicy<C, M, R>

Constructs a fluent builder for the DeleteResourcePolicy operation.

See DeleteResourcePolicy for more information about the operation and its arguments.

pub fn delete_source_credentials(&self) -> DeleteSourceCredentials<C, M, R>

Constructs a fluent builder for the DeleteSourceCredentials operation.

See DeleteSourceCredentials for more information about the operation and its arguments.

pub fn delete_webhook(&self) -> DeleteWebhook<C, M, R>

Constructs a fluent builder for the DeleteWebhook operation.

See DeleteWebhook for more information about the operation and its arguments.

pub fn describe_code_coverages(&self) -> DescribeCodeCoverages<C, M, R>

Constructs a fluent builder for the DescribeCodeCoverages operation.

See DescribeCodeCoverages for more information about the operation and its arguments.

pub fn describe_test_cases(&self) -> DescribeTestCases<C, M, R>

Constructs a fluent builder for the DescribeTestCases operation.

See DescribeTestCases for more information about the operation and its arguments.

pub fn get_report_group_trend(&self) -> GetReportGroupTrend<C, M, R>

Constructs a fluent builder for the GetReportGroupTrend operation.

See GetReportGroupTrend for more information about the operation and its arguments.

pub fn get_resource_policy(&self) -> GetResourcePolicy<C, M, R>

Constructs a fluent builder for the GetResourcePolicy operation.

See GetResourcePolicy for more information about the operation and its arguments.

pub fn import_source_credentials(&self) -> ImportSourceCredentials<C, M, R>

Constructs a fluent builder for the ImportSourceCredentials operation.

See ImportSourceCredentials for more information about the operation and its arguments.

pub fn invalidate_project_cache(&self) -> InvalidateProjectCache<C, M, R>

Constructs a fluent builder for the InvalidateProjectCache operation.

See InvalidateProjectCache for more information about the operation and its arguments.

pub fn list_build_batches(&self) -> ListBuildBatches<C, M, R>

Constructs a fluent builder for the ListBuildBatches operation.

See ListBuildBatches for more information about the operation and its arguments.

pub fn list_build_batches_for_project(
&self
) -> ListBuildBatchesForProject<C, M, R>

Constructs a fluent builder for the ListBuildBatchesForProject operation.

See ListBuildBatchesForProject for more information about the operation and its arguments.

pub fn list_builds(&self) -> ListBuilds<C, M, R>

Constructs a fluent builder for the ListBuilds operation.

See ListBuilds for more information about the operation and its arguments.

pub fn list_builds_for_project(&self) -> ListBuildsForProject<C, M, R>

Constructs a fluent builder for the ListBuildsForProject operation.

See ListBuildsForProject for more information about the operation and its arguments.

pub fn list_curated_environment_images(
&self
) -> ListCuratedEnvironmentImages<C, M, R>

Constructs a fluent builder for the ListCuratedEnvironmentImages operation.

See ListCuratedEnvironmentImages for more information about the operation and its arguments.

pub fn list_projects(&self) -> ListProjects<C, M, R>

Constructs a fluent builder for the ListProjects operation.

See ListProjects for more information about the operation and its arguments.

pub fn list_report_groups(&self) -> ListReportGroups<C, M, R>

Constructs a fluent builder for the ListReportGroups operation.

See ListReportGroups for more information about the operation and its arguments.

pub fn list_reports(&self) -> ListReports<C, M, R>

Constructs a fluent builder for the ListReports operation.

See ListReports for more information about the operation and its arguments.

pub fn list_reports_for_report_group(
&self
) -> ListReportsForReportGroup<C, M, R>

Constructs a fluent builder for the ListReportsForReportGroup operation.

See ListReportsForReportGroup for more information about the operation and its arguments.

pub fn list_shared_projects(&self) -> ListSharedProjects<C, M, R>

Constructs a fluent builder for the ListSharedProjects operation.

See ListSharedProjects for more information about the operation and its arguments.

pub fn list_shared_report_groups(&self) -> ListSharedReportGroups<C, M, R>

Constructs a fluent builder for the ListSharedReportGroups operation.

See ListSharedReportGroups for more information about the operation and its arguments.

pub fn list_source_credentials(&self) -> ListSourceCredentials<C, M, R>

Constructs a fluent builder for the ListSourceCredentials operation.

See ListSourceCredentials for more information about the operation and its arguments.

pub fn put_resource_policy(&self) -> PutResourcePolicy<C, M, R>

Constructs a fluent builder for the PutResourcePolicy operation.

See PutResourcePolicy for more information about the operation and its arguments.

pub fn retry_build(&self) -> RetryBuild<C, M, R>

Constructs a fluent builder for the RetryBuild operation.

See RetryBuild for more information about the operation and its arguments.

pub fn retry_build_batch(&self) -> RetryBuildBatch<C, M, R>

Constructs a fluent builder for the RetryBuildBatch operation.

See RetryBuildBatch for more information about the operation and its arguments.

pub fn start_build(&self) -> StartBuild<C, M, R>

Constructs a fluent builder for the StartBuild operation.

See StartBuild for more information about the operation and its arguments.

pub fn start_build_batch(&self) -> StartBuildBatch<C, M, R>

Constructs a fluent builder for the StartBuildBatch operation.

See StartBuildBatch for more information about the operation and its arguments.

pub fn stop_build(&self) -> StopBuild<C, M, R>

Constructs a fluent builder for the StopBuild operation.

See StopBuild for more information about the operation and its arguments.

pub fn stop_build_batch(&self) -> StopBuildBatch<C, M, R>

Constructs a fluent builder for the StopBuildBatch operation.

See StopBuildBatch for more information about the operation and its arguments.

pub fn update_project(&self) -> UpdateProject<C, M, R>

Constructs a fluent builder for the UpdateProject operation.

See UpdateProject for more information about the operation and its arguments.

pub fn update_project_visibility(&self) -> UpdateProjectVisibility<C, M, R>

Constructs a fluent builder for the UpdateProjectVisibility operation.

See UpdateProjectVisibility for more information about the operation and its arguments.

pub fn update_report_group(&self) -> UpdateReportGroup<C, M, R>

Constructs a fluent builder for the UpdateReportGroup operation.

See UpdateReportGroup for more information about the operation and its arguments.

pub fn update_webhook(&self) -> UpdateWebhook<C, M, R>

Constructs a fluent builder for the UpdateWebhook operation.

See UpdateWebhook for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.